minor changes + fix problems with clustering

+++ synthetic_data/eval_comparison.R
@@ -4,13 +4,14 @@ dataset_id = args[1]  # either "generative_model/all_fits/" or "SigFitTest"
 devtools::load_all("~/GitHub/simbascule/")
 devtools::load_all("~/GitHub/bascule/")
 
+# dataset_id = "generative_model/all_fits/"
 # run_id = "matched.2011.compare"
 run_id = "matched.2011.compare_LAST"
 
 out_id = paste0(run_id, ".", dataset_id %>% stringr::str_remove_all("/all_fits/"))
 
-# main_path = "~/Dropbox/dropbox_shared/2022. Basilica/simulations/"
-main_path = "/orfeo/cephfs/scratch/cdslab/ebusca00/signatures/"
+main_path = "~/Dropbox/dropbox_shared/2022. Basilica/simulations/"
+# main_path = "/orfeo/cephfs/scratch/cdslab/ebusca00/signatures/"
 
 save_path = file.path(main_path, "stats_dataframes/")
 
@@ -19,7 +20,7 @@ source("~/GitHub/bascule_validation/synthetic_data/aux_fns/plots_aux_fns.R")
 
 # Generate stats dataframe ##### 
 runids = c("BASCULE", "SigProfiler", "SparseSignatures", "SignatureToolsLib")
-fitnames = c("fit.0", "sigprofiler", "sparsesignatures", "signaturetoolslib")
+fitnames = c("x.fit0.auto", "sigprofiler", "sparsesignatures", "signaturetoolslib")
 
 # path = paste0("~/Dropbox/dropbox_shared/2022. Basilica/simulations/fits_generative_model/all_fits/fits_dn.", run_id, "/")
 # path = file.path(main_path, paste0("fits_generative_model/all_fits/fits_dn.", run_id, "/"))
@@ -30,16 +31,20 @@ cli::cli_text("Files path: {path}\n
 
 stopifnot(all(file.exists(c(path, save_path))))
 
-files = list.files(path, full.names=T, pattern=".Rds")
-
-library(parallel)
-n_cores = detectCores()
-all_stats = mclapply(files, function(fname) {
-  stats_single_data(fname, names_fits=fitnames %>% setNames(runids))
-}, mc.cores=n_cores/4) %>% dplyr::bind_rows()
-
-saveRDS(all_stats, paste0(save_path, "stats_", out_id, ".Rds"))
-
+if (!file.exists(paste0(save_path, "stats_", out_id, ".Rds"))) {
+  files = list.files(path, full.names=T, pattern=glob2rx("simul_fit*.Rds"))
+  # library(parallel)
+  # n_cores = detectCores()
+  # fn = mclappy
+  fn = lapply
+  all_stats = fn(files, function(fname) {
+    stats_single_data(fname, names_fits=fitnames %>% setNames(runids))
+  # }, mc.cores=n_cores/4
+  }
+  ) %>% dplyr::bind_rows()
+  
+  saveRDS(all_stats, paste0(save_path, "stats_", out_id, ".Rds"))
+}
 
 
 # Make plots #####
@@ -47,7 +52,7 @@ all_stats = readRDS(paste0(save_path, "stats_", out_id, ".Rds")) %>%
   compute_quantiles(colname="K_true") %>% 
   dplyr::filter(type=="SBS")
 
-pal = RColorBrewer::brewer.pal(3, name="Dark2")
+pal = RColorBrewer::brewer.pal(length(runids), name="Dark2")
 
 plot_list = list()
 plot_list[["K"]] = all_stats %>%
